Examples in .sb2 format only
Now that the GoPiGo3 extension can be registered with the
Scratch 2 Offline Editor it is more handy to provide
examples in .sb2 format. The Scratch 2 Offline Editor can
easily read and write files with that filename extension,
and the Scratch extension is automatically loaded when
loading a project referencing the extension.

ScratchX can read .sb2 files. The Scratch extension needs
to loaded explicitly though (as before).